Move serde_ibytes to a seperate file
